+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 v1.0.0
 ------
 
-Initial release:
+Releases:
 
-* 
-* 
+* 1.0.1:
+    * Add password confirmation automatic update (view) when the confirmation input value is changed.
+    * Don't display *uniqueness* error message (status:0) but log it instead.